+++ src/main/java/qna/domain/User.java
@@ -8,7 +8,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
-import qna.UnAuthorizedException;
+import qna.exception.UnAuthorizedException;
 
 @Entity
 public class User extends BaseEntityTime {
